sku lookup changes
