+++ src/main.tsx
@@ -13,8 +13,8 @@ import Chart from "./components/ChartIndex";
 import DcOpenOrders from "./components/DcOpenOrders";
 import HomePage from "./components/homePage";
 import OrderDetails from "./components/orderDetails/OrderDetails.js";
-import SKULookup from "./components/skuLookup";
 import Sessions from "./views/sessions";
+import SKULookup from "./views/skuLookup";
 import { ROUTES } from "./components/utils/Utils";
 import "./index.css";
 
@@ -60,7 +60,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
             }
           />
           <Route
-            path={"/sku-lookup"}
+            path={ROUTES.skuLookup}
             element={<App appContent={<SKULookup />} />}
           />
           <Route
